+++ CHANGELOG.md
@@ -37,9 +37,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 ## Unreleased
 
-### State Machine Breaking
+### API Breaking
 
-* [#228](https://github.com/babylonlabs-io/babylon/pull/228) Add inclusion height to unbonding info
+* [#228](https://github.com/babylonlabs-io/babylon/pull/228) Add inclusion height to early unbonding event
 
 ## v0.14.0